requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
httpx==0.24.1
anthropic==0.20.0
pytest==7.4.0
gunicorn==21.2.0
flask-wtf==1.1.1
spacy==3.7.2
numpy==1.26.3
sentence-transformers==2.2.2
huggingface_hub==0.17.3
transformers==4.35.2
torch>=1.6.0
scikit-learn>=1.0.0
redis==5.0.1